+++ src/shared/lib/animatedNavigate/index.ts
@@ -1,1 +1,2 @@
 export { useAnimatedNavigate } from './useAnimatedNavigate';
+export { useBrowserNavigationHandler } from './useBrowserNavigationHandler';